fix exercise 1.21
did not count the tabstops in the original string as the actual amount
of whitespaces but instead as the tabstop width. by checking the cursor
up until the next tabstop with `TABSTOP - ((c + w) % TABSTOP);` we can
tell how much we need to forward the position as.

also adds a python testcase. run with:
```
k-r$ nosetests test_1_21.py
.
----------------------------------------------------------------------
Ran 1 test in 0.446s

OK
```